fix: footer and auth
