+++ src/App.tsx
@@ -1,8 +1,27 @@
 import Header from './components/Header/Header'
 import Footer from './components/Footer/Footer'
 import AppRouter from './utils/AppRouter'
+import { getAuth, onAuthStateChanged } from 'firebase/auth'
+import { setUser } from './store/slices/userSlice'
+import { useAppDispatch } from './hooks/reduxHooks'
 
 function App() {
+  const auth = getAuth()
+  const dispatch = useAppDispatch()
+
+  onAuthStateChanged(auth, (user) => {
+    if (user) {
+      dispatch(
+        setUser({
+          email: user.email,
+          token: user.refreshToken,
+          id: user.uid,
+          name: user.displayName,
+        }),
+      )
+    }
+  })
+
   return (
     <>
       <Header />
