+++ src/components/userForms/SignIn.tsx
@@ -51,7 +51,7 @@ const SignIn: React.FC = () => {
 
       navigate('/')
       dispatch(closeModal())
-      localStorage.setItem('login', `${user.email}`)
+      localStorage.setItem('login', `${user.uid}`)
       location.reload()
     } catch (error) {
       if (error instanceof Error) console.error(error.message)
