+++ src/styles/global.scss
@@ -29,7 +29,6 @@ body {
   min-height: 100vh;
 
   margin: 0 auto;
-  max-width: 1400px;
 }
 
 .page {
@@ -38,7 +37,7 @@ body {
 }
 
 .container {
-  max-width: 1600px;
+  max-width: 1400px;
   margin: 0 auto;
   padding: 15px;
 }
